requirements.txt
chromadb
Django
djangorestframework
djangorestframework-simplejwt
langchain
langchain-chroma
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langsmith
langgraph
wikipedia
mediawikiapi
wikibase-rest-api-client
duckduckgo-search
pypdf